refactor: move controls to separate files
